backport of commit a35acdb84e2445930598663bb8cbc133e9832b4d
